% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayWOE.R
\name{displayWOE}
\alias{displayWOE}
\title{displayWOE}
\usage{
displayWOE(dset, col = "xyz", resp = "y", adjFactor = 0.5, bins = 10)
}
\arguments{
\item{dset}{The data frame containing the data set}

\item{col}{A character respresenting the  name of the attribute . The attribute can either be numeric or categorical}

\item{resp}{A character respresenting the  name of the binary outcome variable
The binary outcome variable may be a factor with two levels or an integer (or numeric ) with two unique values}

\item{adjFactor}{A number or a decimal denoting what is to be added to the number of responses (binary outcome variable is 1 ) or to the number of non responses (binary outcome variable is 0) if either is zero for any level of the attribute}

\item{bins}{A number denoting the number of bins.Default value is 10}
}
\description{
This function displays the Weight of Evidence of the levels of an attribute.
}
\examples{

# Load the German_Credit data set supplied with this package

data("German_Credit")

displayWOE(German_Credit,col="Credit_History",resp="Good_Bad")

}
