% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVCalc.R
\name{IVCalc}
\alias{IVCalc}
\title{IVCalc}
\usage{
IVCalc(dset, resp = "y", bins = 10, adjFactor = 0.5)
}
\arguments{
\item{dset}{The data frame containing the data set}

\item{resp}{A character respresenting the  name of the binary outcome variable
The binary outcome variable may be a factor with two levels or an integer (or numeric ) with two unique values}

\item{bins}{A number denoting the number of bins.Default value is 10}

\item{adjFactor}{A number or a decimal denoting what is to be added to the number of responses (binary outcome variable is 1 ) or to the number of non responses (binary outcome variable is 0) if either is zero for any level of the attribute}
}
\value{
A list containing the tables of Information Values  by levels for every attribute
}
\description{
This function displays the Information Values  by the levels of an attribute
              This information is displayed for all attributes in the data set
}
\examples{

# Load the German_Credit data set supplied with this package

data("German_Credit")

l<-list()

# Call the function as follows

l<-IVCalc(German_Credit,resp="Good_Bad",bins=10)

# Information Value for  the attribute Account_Balance in the German_Credit data

l$Account_Balance


}
