% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelsCollapser.R
\name{levelsCollapser}
\alias{levelsCollapser}
\title{levelsCollapser}
\usage{
levelsCollapser(dset, resp = "y", bins = 10)
}
\arguments{
\item{dset}{The data frame containing the data set}

\item{resp}{A character respresenting the  name of the binary outcome variable
The binary outcome variable may be a factor with two levels or an integer (or numeric ) with two unique values}

\item{bins}{A number denoting the number of bins.Default value is 10}
}
\value{
A list containing the tables of response rate by levels for every attribute
}
\description{
This function displays the response rates by the levels of an attribute
              Levels with similar response rates may be combined
}
\examples{

# Load the German_Credit data set supplied with this package

data("German_Credit")

# Create an empty list

l<-list()

# Call the function as follows

l<-levelsCollapser(German_Credit,resp="Good_Bad",bins=10)

# response rate by levels of the Account_Balance in the German_Credit data

l$Account_Balance

# Collapse levels with similar response percentages.

}
